requirements.txt
django
psycopg2-binary
django-cors-headers
python-dotenv
openai-agents
rich
markdown
pandas